requirements.txt
asgiref==3.3.4
chardet==4.0.0
Django==3.2
django-environ==0.4.5
djangorestframework==3.12.4
psycopg2-binary==2.8.6
pysub-parser==1.4.1
pytz==2021.1
sqlparse==0.4.1
testfixtures==6.17.1
Unidecode==1.2.0
pylint==2.7.0
pylint_django==2.3.0
Pycco==0.6.0
gunicorn==20.1.0